+++ figlet4s-core/src/main/scala/com/colofabrix/scala/figlet4s/core/FontListing.scala
@@ -5,6 +5,7 @@ import cats.implicits._
 import com.colofabrix.scala.figlet4s.errors._
 import java.io._
 import java.net._
+import java.nio.file.{ Files, Paths }
 import java.util.zip._
 
 private[figlet4s] object FontListing {
@@ -26,7 +27,7 @@ private[figlet4s] object FontListing {
     val file = new File(resources)
 
     if (file.isDirectory)
-      internalFontsFromDirectory(resources).map(_.toSeq)
+      internalFontsFromDirectory(resources.resolve("fonts")).map(_.toSeq)
     else if (file.getName.toLowerCase.endsWith(".jar"))
       internalFontsFromJar(resources).map(_.toSeq)
     else
@@ -37,18 +38,27 @@ private[figlet4s] object FontListing {
 
   //  Support  //
 
-  private def dirReaderStream(resources: URI): BufferedReader =
-    new BufferedReader(new InputStreamReader(resources.resolve("fonts").toURL.openStream()))
+  private def uriBufferedReader(resources: URI): BufferedReader =
+    new BufferedReader(new InputStreamReader(resources.toURL.openStream()))
 
+  @SuppressWarnings(Array("org.wartremover.warts.All"))
   private def internalFontsFromDirectory[F[_]: Sync](resources: URI): F[List[String]] =
-    Braket.withResource(dirReaderStream(resources)) { reader =>
-      Sync[F].delay {
+    Braket.withResource(uriBufferedReader(resources)) { reader =>
+      Sync[F].defer {
         Iterator
           .continually(reader.readLine)
           .takeWhile(Option(_).isDefined)
-          .withFilter(path => path.endsWith(".flf"))
-          .map(_.replace(".flf", ""))
           .toList
+          .traverse { name =>
+            val path = Paths.get(resources.getRawPath(), name)
+            if (Files.isDirectory(path)) internalFontsFromDirectory(path.toUri())
+            else Sync[F].pure(List(path.toString()))
+          }
+          .map {
+            _.flatten
+              .withFilter(path => path.endsWith(".flf"))
+              .map(_.replace(".flf", ""))
+          }
       }
     }